+++ x/auth/vesting/client/cli/tx_test.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"io"
 	"testing"
-	"time"
 
 	rpcclientmock "github.com/cometbft/cometbft/rpc/client/mock"
 	"github.com/stretchr/testify/suite"
@@ -52,174 +51,6 @@ func (s *CLITestSuite) SetupSuite() {
 		WithConsensusAddressCodec(addresscodec.NewBech32Codec("cosmosvalcons"))
 }
 
-func (s *CLITestSuite) TestNewMsgCreateVestingAccountCmd() {
-	accounts := testutil.CreateKeyringAccounts(s.T(), s.kr, 1)
-	cmd := cli.NewMsgCreateVestingAccountCmd()
-	cmd.SetOutput(io.Discard)
-
-	extraArgs := []string{
-		fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastSync),
-		fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
-		fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin("photon", sdkmath.NewInt(10))).String()),
-		fmt.Sprintf("--%s=test-chain", flags.FlagChainID),
-		fmt.Sprintf("--%s=%s", flags.FlagFrom, accounts[0].Address),
-	}
-
-	t := time.Date(2033, time.April, 1, 12, 34, 56, 789, time.UTC).Unix()
-
-	testCases := []struct {
-		name      string
-		ctxGen    func() client.Context
-		from, to  sdk.AccAddress
-		amount    sdk.Coins
-		endTime   int64
-		extraArgs []string
-		expectErr bool
-	}{
-		{
-			"valid transaction",
-			func() client.Context {
-				return s.baseCtx
-			},
-			accounts[0].Address,
-			accounts[0].Address,
-			sdk.NewCoins(
-				sdk.NewCoin("stake", sdkmath.NewInt(10)),
-				sdk.NewCoin("photon", sdkmath.NewInt(40)),
-			),
-			t,
-			extraArgs,
-			false,
-		},
-		{
-			"invalid to Address",
-			func() client.Context {
-				return s.baseCtx
-			},
-			accounts[0].Address,
-			sdk.AccAddress{},
-			sdk.NewCoins(
-				sdk.NewCoin("stake", sdkmath.NewInt(10)),
-				sdk.NewCoin("photon", sdkmath.NewInt(40)),
-			),
-			t,
-			extraArgs,
-			true,
-		},
-		{
-			"invalid coins",
-			func() client.Context {
-				return s.baseCtx
-			},
-			accounts[0].Address,
-			accounts[0].Address,
-			nil,
-			t,
-			extraArgs,
-			true,
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-		s.Run(tc.name, func() {
-			ctx := svrcmd.CreateExecuteContext(context.Background())
-
-			cmd.SetContext(ctx)
-			fmt.Println(tc.amount.String())
-			cmd.SetArgs(append([]string{tc.to.String(), tc.amount.String(), fmt.Sprint(tc.endTime)}, tc.extraArgs...))
-
-			s.Require().NoError(client.SetCmdClientContextHandler(tc.ctxGen(), cmd))
-
-			err := cmd.Execute()
-			if tc.expectErr {
-				s.Require().Error(err)
-			} else {
-				s.Require().NoError(err)
-			}
-		})
-	}
-}
-
-func (s *CLITestSuite) TestNewMsgCreatePermanentLockedAccountCmd() {
-	accounts := testutil.CreateKeyringAccounts(s.T(), s.kr, 1)
-	cmd := cli.NewMsgCreatePermanentLockedAccountCmd()
-	cmd.SetOutput(io.Discard)
-
-	extraArgs := []string{
-		fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastSync),
-		fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
-		fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin("photon", sdkmath.NewInt(10))).String()),
-		fmt.Sprintf("--%s=test-chain", flags.FlagChainID),
-		fmt.Sprintf("--%s=%s", flags.FlagFrom, accounts[0].Address),
-	}
-
-	testCases := []struct {
-		name      string
-		ctxGen    func() client.Context
-		to        sdk.AccAddress
-		amount    sdk.Coins
-		extraArgs []string
-		expectErr bool
-	}{
-		{
-			"valid transaction",
-			func() client.Context {
-				return s.baseCtx
-			},
-			accounts[0].Address,
-			sdk.NewCoins(
-				sdk.NewCoin("stake", sdkmath.NewInt(10)),
-				sdk.NewCoin("photon", sdkmath.NewInt(40)),
-			),
-			extraArgs,
-			false,
-		},
-		{
-			"invalid to Address",
-			func() client.Context {
-				return s.baseCtx
-			},
-			sdk.AccAddress{},
-			sdk.NewCoins(
-				sdk.NewCoin("stake", sdkmath.NewInt(10)),
-				sdk.NewCoin("photon", sdkmath.NewInt(40)),
-			),
-			extraArgs,
-			true,
-		},
-		{
-			"invalid coins",
-			func() client.Context {
-				return s.baseCtx
-			},
-			accounts[0].Address,
-			nil,
-			extraArgs,
-			true,
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-		s.Run(tc.name, func() {
-			ctx := svrcmd.CreateExecuteContext(context.Background())
-
-			cmd.SetContext(ctx)
-			cmd.SetArgs(append([]string{tc.to.String(), tc.amount.String()}, tc.extraArgs...))
-
-			s.Require().NoError(client.SetCmdClientContextHandler(tc.ctxGen(), cmd))
-
-			err := cmd.Execute()
-			if tc.expectErr {
-				s.Require().Error(err)
-			} else {
-				s.Require().NoError(err)
-			}
-		})
-	}
-}
-
 func (s *CLITestSuite) TestNewMsgCreatePeriodicVestingAccountCmd() {
 	accounts := testutil.CreateKeyringAccounts(s.T(), s.kr, 1)
 	cmd := cli.NewMsgCreatePeriodicVestingAccountCmd()
@@ -234,29 +65,22 @@ func (s *CLITestSuite) TestNewMsgCreatePeriodicVestingAccountCmd() {
 	}
 
 	testCases := []struct {
-		name      string
-		ctxGen    func() client.Context
-		to        sdk.AccAddress
-		extraArgs []string
-		expectErr bool
+		name         string
+		to           sdk.AccAddress
+		extraArgs    []string
+		expectErrMsg string
 	}{
 		{
 			"valid transaction",
-			func() client.Context {
-				return s.baseCtx
-			},
 			accounts[0].Address,
 			extraArgs,
-			false,
+			"",
 		},
 		{
-			"invalid to Address",
-			func() client.Context {
-				return s.baseCtx
-			},
+			"invalid to address",
 			sdk.AccAddress{},
 			extraArgs,
-			true,
+			"empty address string is not allowed",
 		},
 	}
 
@@ -266,13 +90,13 @@ func (s *CLITestSuite) TestNewMsgCreatePeriodicVestingAccountCmd() {
 			ctx := svrcmd.CreateExecuteContext(context.Background())
 
 			cmd.SetContext(ctx)
-			cmd.SetArgs(append([]string{tc.to.String(), "./test.json"}, tc.extraArgs...))
+			cmd.SetArgs(append([]string{tc.to.String(), "./periods.json"}, tc.extraArgs...))
 
-			s.Require().NoError(client.SetCmdClientContextHandler(tc.ctxGen(), cmd))
+			s.Require().NoError(client.SetCmdClientContextHandler(s.baseCtx, cmd))
 
 			err := cmd.Execute()
-			if tc.expectErr {
-				s.Require().Error(err)
+			if tc.expectErrMsg != "" {
+				s.Require().ErrorContains(err, "empty address string is not allowed")
 			} else {
 				s.Require().NoError(err)
 			}